refactor(streaming): move stateful hooks to src/hooks/streaming/
`useFrameRenderer`, `useInputLoop`, `usePauseState`, `usePointerLock`
are React hooks, not context modules — they own internal state and
expose imperative actions. Move them under `src/hooks/streaming/`
alongside the rest of the project's hooks.

`src/context/streaming/` keeps just the context plumbing — the eight
`Context + use*Context + value-type` modules and the `sessionConfig`
helper that constructs the wire payload.
