+++ src/context/StreamingContext.tsx
@@ -22,10 +22,10 @@ import useSeedsDir from '../hooks/useSeedsDir'
 import { invoke } from '../bridge'
 import { createLogger } from '../utils/logger'
 import { buildSessionConfig } from './streaming/sessionConfig'
-import { useFrameRenderer } from './streaming/useFrameRenderer'
-import { useInputLoop } from './streaming/useInputLoop'
-import { usePauseState } from './streaming/usePauseState'
-import { usePointerLock } from './streaming/usePointerLock'
+import { useFrameRenderer } from '../hooks/streaming/useFrameRenderer'
+import { useInputLoop } from '../hooks/streaming/useInputLoop'
+import { usePauseState } from '../hooks/streaming/usePauseState'
+import { usePointerLock } from '../hooks/streaming/usePointerLock'
 import { ConnectionContext, type ConnectionContextValue } from './streaming/connection'
 import { EngineContext, type EngineContextValue } from './streaming/engine'
 import { SessionContext, type SessionContextValue } from './streaming/session'
