+++ src/context/streaming/session.ts
@@ -1,6 +1,6 @@
 import type { SceneEditState, SceneEditEvent } from '../sceneEditMachine'
 import { createStreamingContext } from './createStreamingContext'
-import type { PauseState } from './usePauseState'
+import type { PauseState } from '../../hooks/streaming/usePauseState'
 
 /** Pause / scene-edit / menu lifecycle state for the active session. */
 export type SessionContextValue = {
