+++ src/hooks/streaming/useFrameRenderer.ts
