+++ src/hooks/streaming/useInputLoop.ts
