+++ src/hooks/streaming/usePauseState.ts
