+++ src/hooks/streaming/usePointerLock.ts
